Cleanup, cases grouping, links and paths in order
Since integration testing is supposed to test the connectivity between
different modules, the necessary parts were left. The code checks
whether the API endpoints are available for a call, then it reads all
necessary files for testing, then it calls objects for maing calls to
the API, then it writes a new Blueprint to the file system ().