Fix bug with recursion in tests
The newly added functionality to exists() to select the newest resource for SDC
would start checking object versions to return the latest one eventually.
Some tests didn't have versions specified, so the method exists() would call other
methods that would make an HTTP request. And with a not found version
in a response the execution of a programm would run in circle going back to exists()
again.